Cleaning up ruby script options.

+++ rins_ccls/add_descriptions_to_results.rb
@@ -5,37 +5,37 @@ require 'optparse'
 # This hash will hold all of the options
 # parsed from the command-line by
 # OptionParser.
-options = {}
+options = {
+	:db      => "/Volumes/cube/working/indexes/nt",
+	:verbose => false
+}
 
 optparse = OptionParser.new do |opts|
 	# Set a banner, displayed at the top of the help screen.
-	opts.banner = "Usage: #{$0} [options] -i in_file -o out_file ..."
+	opts.banner = "Usage: #{$0} [options] -i in_file -o out_file\n" <<
+		"This script was designed to add the sequence description\n" <<
+				"to the results.txt file produced by the write_result.pl script"
 
-	options[:in] = nil
-	opts.on( '-i', '--in FILE', 'tab delimited txt infile' ) do |file|
+	#	Define the options, and what they do
+
+	opts.on( '-i', '--in FILE', 'tab delimited txt infile (required)' ) do |file|
 		options[:in] = file
 	end
 
-	options[:out] = nil
-	opts.on( '-o', '--out FILE', 'tab delimited txt outfile' ) do |file|
+	opts.on( '-o', '--out FILE', 'tab delimited txt outfile (required)' ) do |file|
 		options[:out] = file
 	end
 
-	# Define the options, and what they do
+	opts.on( '-d', '--db BlastDB', "Blast database reference (default #{options[:db]})" ) do |db|
+		options[:db] = db
+	end
 
-#	options[:suffix] = 'duplicate'
-#	opts.on( '-s', '--suffix STRING', 'Append duplicate sequence name with STRING' ) do |s|
-#		options[:suffix] = s
-#	end
-
-	options[:verbose] = false
 	opts.on( '-v', '--verbose', 'Output more information' ) do
 		options[:verbose] = true
 	end
 
-	# This displays the help screen, all programs are
-	# assumed to have this option.
-	opts.on( '-h', '--help', 'Display this screen' ) do
+	# This displays the help screen, all programs are assumed to have this option.
+	opts.on( '-h', '--help', 'Display this help screen') do
 		puts opts
 		exit
 	end
@@ -47,6 +47,17 @@ end
 # any options found there, as well as any parameters for
 # the options. What's left is the list of files to resize.
 optparse.parse!
+
+#	Infile and outfile required
+unless options[:in] and options[:out]
+	puts optparse	#	Basically display the command line help
+	exit
+end
+
+unless File.exists?(options[:in])
+	puts "File #{options[:in]} not found." 
+	exit
+end
  
 File.open(  options[:in], 'r' ) { |infile|
 File.open( options[:out], 'w' ) { |outfile|
@@ -64,17 +75,14 @@ File.open( options[:out], 'w' ) { |outfile|
 
 		puts "Initial line without description ...\n#{line}" if options[:verbose]
 		parts = line.split("\t")
-		command = "blastdbcmd -entry '#{parts[2]}' -db /Volumes/cube/working/indexes/nt | head -1"
+		command = "blastdbcmd -entry '#{parts[2]}' -db #{options[:db]} | head -1"
 		puts "Executing ...\n#{command}" if options[:verbose]
 		description = `#{command}`
 
 		puts "Got ...\n#{description}" if options[:verbose]
 
 		description.gsub!(">#{parts[2]}",'')
-#		description.gsub!(/^>#{parts[2]}\s*/,'')
-#		description.gsub!(/^>/,'')
 		description.gsub!(/^\s*/,'')
-#		description.gsub!(parts[2],'')
 		puts "After parsing out seqid ...\n#{description}" if options[:verbose]
 
 #> blastdbcmd -entry "gi|301795361|emb|FQ312042.1|" -db /Volumes/cube/working/indexes/nt | head -1
